kvcoord: move TestPartialPartition to own file
